DLL loader updated from avifile-0.60beta4
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@129 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ loader/wine/windef.h
@@ -113,7 +113,7 @@ extern "C" {
 #  define __RESTORE_ES  __asm__ __volatile__("pushl %ds\n\tpopl %es")
 #  endif
 # else
-//#  error You need gcc >= 2.7 to build Wine on a 386
+#  error You need gcc >= 2.7 to build Wine on a 386
 # endif
 #else 
 # define __stdcall
@@ -166,8 +166,8 @@ typedef unsigned short  WCHAR;
 typedef int             WIN_BOOL;
 typedef double          DATE;
 typedef double          DOUBLE;
-typedef double          LONGLONG;
-typedef double          ULONGLONG;
+typedef long long       LONGLONG;
+typedef unsigned long long   ULONGLONG;
 
 /* FIXME: Wine does not compile with strict on, therefore strict
  * handles are presently only usable on machines where sizeof(UINT) ==
